+++ scout/constants/igv_tracks.py
@@ -13,14 +13,21 @@ HG38CYTOBAND_URL = (
     "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/cytoBandIdeo.txt"
 )
 
-# ----- Gencode genes annotation tracks -----
-HG19GENES_FORMAT = "gtf"
-HG19GENES_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg19/genes/gencode.v24lift37.annotation.sorted.gtf.gz"
-HG19GENES_INDEX_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg19/genes/gencode.v24lift37.annotation.sorted.gtf.gz.tbi"
-HG38GENES_FORMAT = "gtf"
-HG38GENES_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/genes/gencode.v24.annotation.sorted.gtf.gz"
-HG38GENES_INDEX_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/genes/gencode.v24.annotation.sorted.gtf.gz.tbi"
+# ----- refSeq genes annotation tracks -----
+HG19GENES_FORMAT = "bed"
+HG19GENES_URL = " https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726131/NCBI_Refseq_genes_GRCh37_sorted.bed.gz"
+HG19GENES_INDEX_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726128/NCBI_Refseq_genes_GRCh37_sorted.bed.gz.tbi"
+HG38GENES_FORMAT = "bed"
+HG38GENES_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726146/NCBI_Refseq_genes_GRCh38_sorted.bed.gz"
+HG38GENES_INDEX_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726143/NCBI_Refseq_genes_GRCh38_sorted.bed.gz.tbi"
 
+# ----- GenCode genes annotation tracks -----
+GENCODE_HG19_GENES_FORMAT = "bed"
+GENCODE_HG19_GENES_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726194/GenCode_comprehensive_genes_v38_lift37_sorted.bed.gz"
+GENCODE_HG19_GENES_INDEX_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726191/GenCode_comprehensive_genes_v38_lift37_sorted.bed.gz.tbi"
+GENCODE_HG38_GENES_FORMAT = "bed"
+GENCODE_HG38_GENES_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726227/GenCode_comprehensive_genes_v38_sorted.bed.gz"
+GENCODE_HG38_GENES_INDEX_URL = "https://s3-eu-west-1.amazonaws.com/pfigshare-u-files/30726224/GenCode_comprehensive_genes_v38_sorted.bed.gz.tbi"
 
 # ----- ClinVar tracks -----
 HG38CLINVAR_URL = "https://hgdownload.soe.ucsc.edu/gbdb/hg38/bbi/clinvar/clinvarMain.bb"
@@ -93,9 +100,10 @@ CLINVAR_CNV_38 = {
     "url": HG38CLINVAR_CNVS_URL,
 }
 
+
 # Human genes track, build 37
-HUMAN_GENES_37 = {
-    "name": "Genes",
+REFSEQ_GENES_37 = {
+    "name": "RefSeq Genes",
     "track_name": "genes_track",
     "type": "annotation",
     "sourceType": "file",
@@ -106,8 +114,8 @@ HUMAN_GENES_37 = {
 }
 
 # Human genes track, build 38
-HUMAN_GENES_38 = {
-    "name": "Genes",
+REFSEQ_GENES_38 = {
+    "name": "RefSeq Genes",
     "track_name": "genes_track",
     "type": "annotation",
     "sourceType": "file",
@@ -115,6 +123,28 @@ HUMAN_GENES_38 = {
     "format": HG38GENES_FORMAT,
     "url": HG38GENES_URL,
     "indexURL": HG38GENES_INDEX_URL,
+}
+
+GENCODE_GENES_37 = {
+    "name": "GenCode Genes",
+    "track_name": "gencode_track",
+    "type": "annotation",
+    "sourceType": "file",
+    "displayMode": "EXPANDED",
+    "format": GENCODE_HG19_GENES_FORMAT,
+    "url": GENCODE_HG19_GENES_URL,
+    "indexURL": GENCODE_HG19_GENES_INDEX_URL,
+}
+
+GENCODE_GENES_38 = {
+    "name": "GenCode Genes",
+    "track_name": "gencode_track",
+    "type": "annotation",
+    "sourceType": "file",
+    "displayMode": "EXPANDED",
+    "format": GENCODE_HG38_GENES_FORMAT,
+    "url": GENCODE_HG38_GENES_URL,
+    "indexURL": GENCODE_HG38_GENES_INDEX_URL,
 }
 
 CASE_SPECIFIC_TRACKS = {
@@ -127,10 +157,10 @@ CASE_SPECIFIC_TRACKS = {
 
 HUMAN_REFERENCE = {"37": HUMAN_REFERENCE_37, "38": HUMAN_REFERENCE_38}
 
-USER_DEFAULT_TRACKS = ["Genes", "ClinVar", "ClinVar CNVs"]
+USER_DEFAULT_TRACKS = ["RefSeq Genes", "GenCode genes", "ClinVar", "ClinVar CNVs"]
 
 # Export selectable custom tracks into lists
 IGV_TRACKS = {
-    "37": [HUMAN_GENES_37, CLINVAR_SNV_37, CLINVAR_CNV_37],
-    "38": [HUMAN_GENES_38, CLINVAR_SNV_38, CLINVAR_CNV_38],
+    "37": [REFSEQ_GENES_37, GENCODE_GENES_37, CLINVAR_SNV_37, CLINVAR_CNV_37],
+    "38": [REFSEQ_GENES_38, GENCODE_GENES_38, CLINVAR_SNV_38, CLINVAR_CNV_38],
 }